remote_caption_mcp_server/requirements.txt
huggingface_hub==0.30.1
transformers~=4.51.3
pillow~=11.1.0
accelerate~=1.7.0
fastmcp>=2.6.1
loguru>=0.7.3
gradio~=5.32.1
#liger_kernel